refactor hh webhook module
