+++ app/api/hh_webhook/__init__.py
@@ -1,0 +1,3 @@
+from .routes import ensure_hh_webhook
+
+__all__ = ["ensure_hh_webhook"]
